Extract implementations of `Contact`s to separate files; correct package name
